Добавит компонент-заглушку на случай отсутствия offers
