+++ project/src/pages/cities-container/cities-container.tsx
@@ -1,0 +1,64 @@
+import { City } from '../../types/offers-types';
+import Map from '../map/map';
+import { MapMode } from '../../const';
+import { Offers, OfferType } from '../../types/offers-types';
+import OffersList from '../offers-list/offers-list';
+import { useState } from 'react';
+
+type CitiesContainerPropsType = {
+  city: City;
+  offers: Offers;
+  placesCount: number;
+};
+
+function CitiesContainer({ city, offers, placesCount }: CitiesContainerPropsType): JSX.Element {
+  const [selectedPoint, setSelectedPoint] = useState<OfferType | undefined>(undefined);
+
+  const onListItemHover = (listItemName: string) => {
+    const currentPoint = offers.find((offer) => offer.title === listItemName);
+
+    setSelectedPoint(currentPoint);
+  };
+
+  return (
+    <div className="cities__places-container container">
+      <section className="cities__places places">
+        <h2 className="visually-hidden">Places</h2>
+        <b className="places__found">{placesCount} places to stay in Amsterdam</b>
+        <form className="places__sorting" action="#" method="get">
+          <span className="places__sorting-caption">Sort by</span>
+          <span className="places__sorting-type" tabIndex={0}>
+            Popular
+            <svg className="places__sorting-arrow" width="7" height="4">
+              <use xlinkHref="#icon-arrow-select"></use>
+            </svg>
+          </span>
+          <ul className="places__options places__options--custom places__options--opened">
+            <li className="places__option places__option--active" tabIndex={0}>
+              Popular
+            </li>
+            <li className="places__option" tabIndex={0}>
+              Price: low to high
+            </li>
+            <li className="places__option" tabIndex={0}>
+              Price: high to low
+            </li>
+            <li className="places__option" tabIndex={0}>
+              Top rated first
+            </li>
+          </ul>
+        </form>
+        <div className="cities__places-list places__list tabs__content">
+          {offers.map((offer) => (
+            <OffersList offer={offer} key={offer.id} onListItemHover={onListItemHover} />
+          ))}
+        </div>
+      </section>
+      <div className="cities__right-section">
+        <Map offers={offers} city={city} selectedPoint={selectedPoint} mapMode={MapMode.Main} />
+      </div>
+    </div>
+  );
+}
+
+export default CitiesContainer;
