+++ project/src/pages/not-found-page/not-found-page.tsx
@@ -1,6 +1,7 @@
 import { Link } from 'react-router-dom';
+import { AppRoute } from '../../const';
 
-function NotFoundPage (): JSX.Element {
+function NotFoundPage(): JSX.Element {
   return (
     <div>
       <header className="header">
@@ -15,8 +16,7 @@ function NotFoundPage (): JSX.Element {
               <ul className="header__nav-list">
                 <li className="header__nav-item user">
                   <a className="header__nav-link header__nav-link--profile" href="/">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
+                    <div className="header__avatar-wrapper user__avatar-wrapper"></div>
                     <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
                   </a>
                 </li>
@@ -34,12 +34,13 @@ function NotFoundPage (): JSX.Element {
       <div className="cities__places-container container">
         <section className="cities__places places">
           <h2 className="visually-hidden">Places</h2>
-          <b className="places__found"><p>404.Страница не найдена</p></b>
-          <Link to="/">Вернуться на главную</Link>
+          <b className="places__found">
+            <p>404.Страница не найдена</p>
+          </b>
+          <Link to={AppRoute.Main}>Вернуться на главную</Link>
         </section>
       </div>
     </div>
-
   );
 }
